backend/requirements.txt
Django>=5.1.7
djangorestframework>=3.15.2
django-cors-headers>=4.7.0
python-dotenv>=1.0.1
mongoengine>=0.27.0
# Testing tools
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
mongomock>=4.1.2 